backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
django==4.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==3.14.0
    # via -r requirements.in
django-rest-framework==0.1.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via django-rest-framework
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
